requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.2
botocore==1.38.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-simple-history==3.8.0
django-storages==1.14.6
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.16.0
djangorestframework-stubs==3.15.3
djangorestframework_simplejwt==5.5.0
idna==3.10
jmespath==1.0.1
mypy==1.15.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
types-stripe==3.5.2.20240106
typing_extensions==4.13.2
urllib3==2.4.0